+++ seed/csharp-sdk/csharp-grpc-proto-exhaustive/include-exception-handler/src/SeedApi/DataService/Requests/UpdateRequest.cs
@@ -14,7 +14,7 @@ public record UpdateRequest
     public required string Id { get; set; }
 
     [JsonPropertyName("values")]
-    public IEnumerable<float>? Values { get; set; }
+    public IEnumerable<double>? Values { get; set; }
 
     [JsonPropertyName("set_metadata")]
     public Metadata? SetMetadata { get; set; }
